Refactor create-package tool and update TypeScript configuration
- Change command for creating packages to use 'tsx' instead of 'node'.
- Add tests for the example package.
- Update tsconfig to exclude tests from compilation.
- Add new dependencies: commander and tsx.
- Remove old create-package package.json and tsconfig.json.
- Update root tsconfig to remove references to the removed create-package.
