+++ package.json
@@ -22,7 +22,7 @@
     "lint": "prettier --check .",
     "format": "prettier --write .",
     "clean": "pnpm -r run clean && rm -rf node_modules",
-    "create-package": "node tools/create-package/dist/index.js",
+    "create-package": "tsx tools/create-package/index.ts",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build"
   },
@@ -32,14 +32,17 @@
     "@storybook/blocks": "^8.0.0",
     "@storybook/react": "^8.0.0",
     "@storybook/react-vite": "^8.0.0",
+    "@types/commander": "^2.12.5",
     "@types/node": "^20.0.0",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "@vitest/ui": "^2.0.0",
+    "commander": "^14.0.1",
     "prettier": "^3.0.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "storybook": "^8.0.0",
+    "tsx": "^4.19.1",
     "typescript": "^5.0.0",
     "vite": "^5.0.0",
     "vite-tsconfig-paths": "^5.1.4",
