--- tools/create-package/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@domeadev/create-package",
-  "version": "1.0.0",
-  "private": true,
-  "description": "CLI tool for creating new @domeadev packages",
-  "type": "module",
-  "bin": {
-    "create-package": "./dist/index.js"
-  },
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc --watch",
-    "clean": "rm -rf dist *.tsbuildinfo"
-  },
-  "dependencies": {
-    "commander": "^12.0.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.0.0",
-    "typescript": "^5.0.0"
-  }
-}
